+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/unlinkaccount/UnlinkAccountModule.kt
@@ -1,15 +1,15 @@
-package io.horizontalsystems.bankwallet.modules.manageaccount.dialogs
+package io.horizontalsystems.bankwallet.modules.unlinkaccount
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.entities.Account
 
-object UnlinkConfirmationDialogModule {
+object UnlinkAccountModule {
     class Factory(private val account: Account) : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return UnlinkConfirmationDialogViewModel(account, App.accountManager) as T
+            return UnlinkAccountViewModel(account, App.accountManager) as T
         }
     }
 }
